+++ framework/src/passes/tile_graph_pass/graph_partition/n_buffer_merge.cpp
@@ -228,9 +228,9 @@ Status NBufferMerge::Init(Function &func) {
     return SUCCESS;
 }
 
-std::map<int, size_t> NBufferMerge::GetIsoColorMergeNum(const OperationsViewer &opOriList,
+std::map<uint64_t, size_t> NBufferMerge::GetIsoColorMergeNum(const OperationsViewer &opOriList,
     const std::map<uint64_t, std::vector<int>> &hashMap) const {
-    std::map<int, size_t> hashCoreNum;
+    std::map<uint64_t, size_t> hashCoreNum;
     for (const auto& entry : hashMap) {
         if (entry.first == 0 || entry.second.empty()) {
             continue;
@@ -403,7 +403,7 @@ void NBufferMerge::MergePingPong(std::vector<std::vector<int>> &sortedColors,
 
 Status NBufferMerge::MergeProcess(const OperationsViewer &opOriList, 
                                       std::map<uint64_t, std::vector<int>> &hashMap, 
-                                      std::map<int, size_t> &hashMergeNum, 
+                                      std::map<uint64_t, size_t> &hashMergeNum, 
                                       std::vector<uint64_t> &hashColor) {
     std::vector<uint64_t> hashMapKeys;
     for (const auto &entry : hashMap) {
@@ -425,17 +425,17 @@ Status NBufferMerge::MergeProcess(const OperationsViewer &opOriList,
     return SUCCESS;
 }
 
-std::map<int, size_t> NBufferMerge::SetNumDB(std::map<uint64_t, std::vector<int>> &hashMap) {
-    std::map<int, size_t> numDBList;
+std::map<uint64_t, size_t> NBufferMerge::SetNumDB(std::map<uint64_t, std::vector<int>> &hashMap) {
+    std::map<uint64_t, size_t> numDBList;
     auto it = vecNBufferSetting.find(VEC_NBUFFER_SETTING_DEFAULT_MERGE_NUM_KEY);
     if (it != vecNBufferSetting.end()) {
         int defaultVal = it->second;
-        for (int i = 0; i < static_cast<int>(hashMap.size()); i++) {
+        for (uint64_t i = 0; i < static_cast<uint64_t>(hashMap.size()); i++) {
             numDBList[i] = defaultVal;
         }
         vecNBufferSetting.erase(it);
     } else {
-        for (int i = 0; i < static_cast<int>(hashMap.size()); i++) {
+        for (uint64_t i = 0; i < static_cast<uint64_t>(hashMap.size()); i++) {
             numDBList[i] = 1;
         }
     }
@@ -467,7 +467,7 @@ Status NBufferMerge::NBufferMergeProcess(Function &func) {
     std::vector<uint64_t> hashColor(color_, 0);
     std::map<uint64_t, std::vector<int>> hashMap;
     GetColorHash(opOriList, hashColor, hashMap);
-    std::map<int, size_t> hashMergeNum;
+    std::map<uint64_t, size_t> hashMergeNum;
     if (vecNBuffermode == 1) {
         if (vecNBufferSetting.size() != 0) {
             APASS_LOG_ERROR_F(Elements::Config, "VEC_NBUFFER_MODE is manually set to 1; Please set VEC_NBUFFER_SETTING to empty.");
