+++ framework/src/passes/tile_graph_pass/graph_partition/n_buffer_merge.h
@@ -44,12 +44,12 @@ private:
     Status CheckAndFixColorOrder(OperationsViewer &opOriList,
                                int &color1, std::vector<int> &colorCycles1,
                                std::vector<std::vector<int>> &colorNode1);
-    std::map<int, size_t> GetIsoColorMergeNum(const OperationsViewer &opOriList,
+    std::map<uint64_t, size_t> GetIsoColorMergeNum(const OperationsViewer &opOriList,
                                                    const std::map<uint64_t, std::vector<int>> &hashMap) const;
     std::vector<std::vector<int>> SortColorWithInput(std::vector<int> &colorValues) const;
     Status MergeProcess(const OperationsViewer &opOriList,
                         std::map<uint64_t, std::vector<int>> &hashMap,
-                        std::map<int, size_t> &hashMergeNum,
+                        std::map<uint64_t, size_t> &hashMergeNum,
                         std::vector<uint64_t> &hashColor);
     Status ColorTopo(int &color1,
                      std::vector<std::vector<int>> &inputColor,
@@ -59,7 +59,7 @@ private:
                        const OperationsViewer &opOriList,
                        std::vector<uint64_t> &hashColor,
                        int &numDBmerge);
-    std::map<int, size_t> SetNumDB(std::map<uint64_t, std::vector<int>> &hashMap);
+    std::map<uint64_t, size_t> SetNumDB(std::map<uint64_t, std::vector<int>> &hashMap);
     Status CheckVecNBufferSettingForManualMerge();
 private:
     int color_{0};
